- add .env-sample file - update the package.json file - remove user.js file to avoid code redudancy - add changePassword function - add changePassword function to helper/user.js - add changePassword route
[delivers #161056968]
