+++ .env-sample
@@ -1,0 +1,14 @@
+PORT=XXXX
+NODE_ENV=XXXX
+DB_HOST=XXX.X.X.X
+DB_PORT=XXXX
+DB_PASSWORD=XXXX
+DB_USERNAME=username of your database
+DB_DATABASE=name of your database
+DB_DATABASE_TEST=name of your test database
+H_PASSWORD=your hash password
+DATABASE_URL=heroku postgres url
+JWT_SECRET='your JWT secret key here'
+FIREBASE_API_KEY=firebase key
+EMAIL_PASS=the app email pass
+DATABASE_LINK=heroku postgres url
