+++ package.json
@@ -31,6 +31,9 @@
   },
   "homepage": "https://github.com/sulenchy/MyDiary#readme",
   "dependencies": {
+    "babel-cli": "^6.26.0",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-stage-2": "^6.24.1",
     "bcrypt-nodejs": "0.0.3",
     "body-parser": "^1.18.3",
     "chai": "^4.1.2",
@@ -43,13 +46,11 @@
     "node-cron": "^1.2.1",
     "node-fetch": "^2.2.0",
     "nodemailer": "^4.6.8",
+    "nodemon": "^1.18.4",
     "pg": "^7.4.3",
     "swagger-ui-express": "^3.0.10",
     "validatorjs": "^3.14.2",
-    "winston": "^3.0.0",
-    "babel-cli": "^6.26.0",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-stage-2": "^6.24.1"
+    "winston": "^3.0.0"
   },
   "devDependencies": {
     "coveralls": "^3.0.2",
