+++ server/controllers/users..js
@@ -190,4 +190,35 @@ export default class UsersController {
           },
         }));
   }
+
+    /**
+     * @description change user password
+     *
+     * @param {*} req
+     * @param {*} res
+     *
+     * @memberOf UsersController Class
+     */
+    static changePassword(req, res) {
+      const email = req.body.email.toLowerCase();
+      const password = req.body.password;
+      usersHelper.changePassword(email, password)
+        .then(user => res.status(200).json({
+          user: {
+            id: user.rows[0].id,
+            email: user.rows[0].email,
+          },
+          message: 'Password changed successfully',
+        }))
+        .catch(err => res.status(500)
+          .json({
+            error: {
+              message: err.message,
+            },
+          }));
+    }
+
+
+
+    
 }
