+++ server/helpers/users.js
@@ -89,6 +89,24 @@ export default class UsersHelper {
     });
   }
 
+  static changePassword(email, password){
+    const hashedPassword = bcrypt.hashSync(`${password}`);
+    const user = `UPDATE users SET password = '${hashedPassword}' WHERE  email = '${email}'  RETURNING *;`;
+    return new Promise((resolve, reject) => {
+      const data = client.query(user);
+      if (data) {
+        resolve(data);
+      } else {
+        reject(new Error({
+          message: 'Sorry, user profile cannot be updated',
+        }));
+      }
+    });
+  }
+
+
+  
+
   static setVisited(userid) {
     const user = `UPDATE users SET visited = NOW()::DATE WHERE id = ${userid}  RETURNING *;`;
     return new Promise((resolve, reject) => {
